producer/requirements.txt
pandas==2.1.4
kafka-python==2.0.2